+++ src/util/minheap.h
@@ -9,13 +9,14 @@
 #define MINHEAP_H_
 
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct {
-    unsigned long db_id; // id of the DB sequence
-    int dframe;          // strand of the DB sequence
-    int dstrand;         // frame the DB sequence
-    int query_id;        // id of the compared query in seq_buffer of search_data
-    long score;
+    size_t db_id;        // id of the DB sequence
+    uint8_t dframe;      // strand of the DB sequence
+    uint8_t dstrand;     // frame the DB sequence
+    uint8_t query_id;    // id of the compared query in seq_buffer of search_data
+    long score;          // score of the alignment
 } elem_t;
 
 typedef struct {
